refactor: extract tpi formula for acte to new file
